Adds an auth that will auth everyone
This is really bad. I based it a bit on the old version of auth, but it
doesn't do much for authorizing since it isn't checking against
anything. Everyone is welcome. This is a proof of concept that we don't
need an auth process. We could use an auth module and have the
CommandHandelr call a function there, but we don't need a server.

Amos King @adkron <amos@binarynoggin.com>