The rest message is runing into the RETR message
I'm not sure how to go from here, but I think the test is good. In order
to check that the test was good I implemented this on the client side of
the FTP to truncate the returned value before passing it out to the
user.

Now that I took out the test hack we need to get the actual server to
handle the messages correctly.

Amos King @adkron <amos@binarynoggin.com>