Development (#25)
* Added option for custom file_handler (#19)

Added an option for a custom file_handler. This will handle
the permissions. If a custom file_handler is not specified, then
Ftp.Permissions will be used.

* Add client_ip_address to struct (#22)

* Add client_ip_address to struct

This allows us to easily access the client (remote) ip address that
is trying to connect to our server.

* switch to full camel case

* fix merge conflicts

* Return atom instead of state in error for stor (#24)

STOR command was returning the state in the tuple which was then
being seen by the user. If you look in the bifrost.erl file, we
can see that upon error for the put_file function, we should be
return 'Info', not 'State'.