Fixes mkdir test never running
The precondition for mkdir was wrong. I was a little worried about there
being a spelling mistake in the state changes too so I changed them to a
module and delegate calls to the module. This should help keeps things
organized and a little safer when we have errors.

Amos King @adkron <amos@binarynoggin.com>